Rename src/encoded/utils.py to util.py
