+++ src/encoded/tests/test_aggregation.py
@@ -1,7 +1,7 @@
 import pytest
 
 from dcicutils.qa_utils import notice_pytest_fixtures
-from ..utils import delay_rerun
+from ..util import delay_rerun
 from .workbook_fixtures import app_settings, app, workbook
