+++ src/encoded/tests/test_batch_download.py
@@ -1,7 +1,7 @@
 import pytest
 
 from dcicutils.qa_utils import notice_pytest_fixtures
-from ..utils import delay_rerun
+from ..util import delay_rerun
 # Use workbook fixture from BDD tests (including elasticsearch)
 from .workbook_fixtures import app_settings, app, workbook
