+++ src/encoded/tests/test_search.py
@@ -11,7 +11,7 @@ from snovault.util import add_default_embeds
 from ..commands.run_upgrader_on_inserts import get_inserts
 # Use workbook fixture from BDD tests (including elasticsearch)
 from .workbook_fixtures import app_settings, app, workbook
-# from ..utils import customized_delay_rerun
+# from ..util import customized_delay_rerun
 
 
 pytestmark = [
