+++ src/encoded/tests/test_types_init_collections.py
@@ -1,7 +1,7 @@
 import pytest
 
 from ..types.image import Image
-from ..utils import utc_today_str
+from ..util import utc_today_str
 
 
 pytestmark = [pytest.mark.setone, pytest.mark.working, pytest.mark.schema]
