+++ src/encoded/tests/test_validation_errors.py
@@ -2,7 +2,7 @@ import pytest
 
 from dcicutils.qa_utils import notice_pytest_fixtures
 from .workbook_fixtures import app_settings, workbook
-# from ..utils import delay_rerun
+# from ..util import delay_rerun
 
 
 notice_pytest_fixtures(app_settings, workbook)
