+++ src/encoded/types/file.py
@@ -63,7 +63,7 @@ from .base import (
     get_item_or_none,
     lab_award_attribution_embed_list
 )
-from ..utils import check_user_is_logged_in
+from ..util import check_user_is_logged_in
 
 
 logging.getLogger('boto3').setLevel(logging.CRITICAL)
